feat: Allegra protocol param updates
* update protocol parameter set from protocol param update spec
* move protocol params Nonce type to common

Fixes #742
